Move DevelopmentOnly to wrappers
- Add styling to only display the lock when hovering the header
